requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
google-api-python-client
google-genai
google_auth_oauthlib
fastapi[standard]